requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]==0.89.0
uvicorn[standard]==0.20.0
gunicorn==20.1.0
tldextract
pytz
google-cloud-bigquery
google-cloud-bigquery-storage
pandas
db-dtypes
uuid
httpx
PyJWT
passlib
bcrypt
asyncio
aiohttp
retrying